requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
mysql-connector-python==8.2.0
playwright==1.40.0
httpx==0.25.2
python-multipart==0.0.6
pydantic==2.5.0
numpy==1.26.4
scikit-learn==1.3.2
python-dotenv==1.0.0
APScheduler==3.10.4
